--- breakdown.html
@@ -1,47 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <title>Basic Breakout HTML Game</title>
-  <meta charset="UTF-8">
-  <link rel="stylesheet" href="breakdown.css">
-</head>
-<body>
-<canvas width="400" height="500" id="game"></canvas>
-<button id="help-button">?</button>
-<div id="help-popup">
-  <p>How to play:</p>
-  <p>Start
-    <ul>
-      Start with pressing the spacebar, if the the ball goes outside of canvas then press space again.
-    </ul>
-  </p>
-  <p>Score
-    <ul>
-      When a ball collides with a brick, the score should increase based on the color. 
-    </ul>
-  </p>  
-  <p> Ball speed
-    <ul>
-      The ball should slowly get faster at certain points in the game
-    </ul>
-  </p> 
-  <p> Sound
-    <ul>
-      Press the sound button if you want there to play sounds
-    </ul>
-  </p> 
-    <button id="close-button">Close</button>
-  </div>
-<button id="play" class = "play-button">🔊</button>
-<script>
-  let play = document.getElementById("play");
-  function playMusic() {
-    let audio = new Audio("breakdown-sound.mp3");
-    audio.play();
-    audio.loop();
-  }
-  play.addEventListener("click", playMusic);
-</script>
-<script src="breakdown.js"></script>
-</body>
-</html>
