--- home.css
@@ -1,164 +1,0 @@
-html, body {
-    overflow: hidden;
-    margin: 0;
-    padding: 0;
-    height: 100%;
-}
-
-body {
-    display: flex;
-    flex-direction: column;
-    background: linear-gradient(180deg, rgba(0,0,0,1) 21%, rgba(98,0,81,1) 63%, rgba(219,0,137,1) 100%);
-}
-
-header {
-    background:transparent;
-    color: #fff;
-    height: 5%;
-}
-
-.hidden {
-    display: none;
-}
-
-
-.logo{
-    margin-left: 2rem;
-    margin-top: 1rem;
-}
-
-.logo-image{
-    position: relative;
-    overflow: hidden;
-    width: 100px;
-    height: 100px;
-}
-  
-.logo-image:hover{
-    background-color: rgba(255, 255, 255, 0.774);
-    border-radius: 10px;
-    box-shadow: 0 0 10px rgba(255, 255, 255, 0.3);
-    backdrop-filter: blur(5px);
-
-    transform: scale(1.05);
-    transition: box-shadow 0.5s ease-in-out, transform 0.5s ease-in-out, background-color 0.5s ease-in-out, backdrop-filter 0.5s ease-in-out;
-
-    -webkit-backdrop-filter: blur(5px);
-    -webkit-transform: scale(1.05);
-
-    cursor: pointer;
-}
-
-
-.main-content {
-    background: transparent;
-    flex: 1;
-    display: flex;
-    flex-direction: column;
-    width: 100vw;
-    height: 85vh;
-}
-
-.hero {
-    height: 60%;
-    width:100vw;
-    display: flex;
-    flex-direction: column;
-    justify-content: top;
-    align-items: center;
-    padding-bottom: 1em;
-}
-
-.hero-text{
-    font-family: 'Roboto', sans-serif;
-    font-size: 4rem;
-    text-align: center;
-}
-
-.gradient-text {
-    background: linear-gradient(90deg, rgb(255, 142, 50), rgb(255, 39, 255));
-    -webkit-background-clip: text;
-    background-clip: text;
-    -webkit-text-fill-color: transparent;
-    -moz-text-fill-color: transparent;
-    -ms-text-fill-color: transparent;
-    -o-text-fill-color: transparent;
-}
-
-.reverse-gradient-text{ 
-    background: linear-gradient(90deg, rgb(255, 39, 255), rgb(255, 142, 50));
-    -webkit-background-clip: text;
-    background-clip: text;
-    -webkit-text-fill-color: transparent;
-    -moz-text-fill-color: transparent;
-    -ms-text-fill-color: transparent;
-    -o-text-fill-color: transparent;
-}
-
-.typewriter p {
-    color: #fff;
-    font-family: 'Roboto', sans-serif;
-    font-size: 2rem;
-    overflow: hidden; /* Ensures the content is not revealed until the animation */
-    border-right: .15em solid rgb(255, 161, 72); /* The typwriter cursor */
-    white-space: nowrap; /* Keeps the content on a single line */
-    margin: 0 auto; /* Gives that scrolling effect as the typing happens */
-    letter-spacing: .15em; /* Adjust as needed */
-    animation: 
-      typing 10s steps(60, end) infinite,
-      blink-caret .5s step-end infinite;
-}
-  
-/* The typing effect */
-@keyframes typing {
-    from, to { width: 0 }
-    50% { width: 100% }
-    70% { width: 100% }
-    95% { width: 0 }
-}
-  
-/* The typewriter cursor effect */
-@keyframes blink-caret {
-    from, to { border-color: transparent }
-    50% { border-color: orange }
-}
-
-.button-wrapper {
-    height: 30%;
-    width: 100vw;
-    display:flex;
-    justify-content: center;
-    align-items: center;
-}
-
-.game-button {
-    background: transparent;
-    display:flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    height: 100%;
-    width: 15vw;
-
-    margin-left: 4rem;
-    margin-right: 4rem;
-
-    cursor:pointer;
-    background-color: rgba(255, 255, 255, 0.781);
-    border-radius: 10px;
-    box-shadow: 0 0 10px rgba(255, 255, 255, 0.3);
-    backdrop-filter: blur(5px);
-
-    transition: all 0.3s ease-in-out;
-}
-
-.game-button:hover{
-
-    box-shadow: 0px 10px 20px rgba(0, 0, 0, 0.3);
-}
-
-.game-text{
-    font-family: 'Roboto', sans-serif;
-    font-size: 2rem;
-    text-align: center;
-}
