+++ index.css
@@ -1,0 +1,312 @@
+html, body {
+    overflow: hidden;
+    margin: 0;
+    padding: 0;
+    height: 100%;
+}
+
+body {
+    display: flex;
+    flex-direction: column;
+    background: linear-gradient(180deg, rgba(0,0,0,1) 21%, rgba(98,0,81,1) 63%, rgba(219,0,137,1) 100%);
+}
+
+header {
+    background:transparent;
+    color: #fff;
+    height: 5%;
+}
+
+.hidden {
+    display: none;
+}
+
+
+.logo{
+    margin-left: 2rem;
+    margin-top: 1rem;
+}
+
+.logo-image{
+    position: relative;
+    overflow: hidden;
+    width: 100px;
+    height: 100px;
+}
+  
+.logo-image:hover{
+    background-color: rgba(255, 255, 255, 0.774);
+    border-radius: 10px;
+    box-shadow: 0 0 10px rgba(255, 255, 255, 0.3);
+    backdrop-filter: blur(5px);
+
+    transform: scale(1.05);
+    transition: box-shadow 0.5s ease-in-out, transform 0.5s ease-in-out, background-color 0.5s ease-in-out, backdrop-filter 0.5s ease-in-out;
+
+    -webkit-backdrop-filter: blur(5px);
+    -webkit-transform: scale(1.05);
+
+    cursor: pointer;
+}
+
+
+.main-content {
+    background: transparent;
+    flex: 1;
+    display: flex;
+    flex-direction: column;
+    width: 100vw;
+    height: 85vh;
+}
+
+.hero {
+    height: 60%;
+    width:100vw;
+    display: flex;
+    flex-direction: column;
+    justify-content: top;
+    align-items: center;
+    padding-bottom: 1em;
+}
+
+.hero-text{
+    font-family: 'Roboto', sans-serif;
+    font-size: 4rem;
+    text-align: center;
+}
+
+.gradient-text {
+    background: linear-gradient(90deg, rgb(255, 142, 50), rgb(255, 39, 255));
+    -webkit-background-clip: text;
+    background-clip: text;
+    -webkit-text-fill-color: transparent;
+    -moz-text-fill-color: transparent;
+    -ms-text-fill-color: transparent;
+    -o-text-fill-color: transparent;
+}
+
+.reverse-gradient-text{ 
+    background: linear-gradient(90deg, rgb(255, 39, 255), rgb(255, 142, 50));
+    -webkit-background-clip: text;
+    background-clip: text;
+    -webkit-text-fill-color: transparent;
+    -moz-text-fill-color: transparent;
+    -ms-text-fill-color: transparent;
+    -o-text-fill-color: transparent;
+}
+
+.typewriter p {
+    color: #fff;
+    font-family: 'Roboto', sans-serif;
+    font-size: 2rem;
+    overflow: hidden; /* Ensures the content is not revealed until the animation */
+    border-right: .15em solid rgb(255, 161, 72); /* The typwriter cursor */
+    white-space: nowrap; /* Keeps the content on a single line */
+    margin: 0 auto; /* Gives that scrolling effect as the typing happens */
+    letter-spacing: .15em; /* Adjust as needed */
+    animation: 
+      typing 10s steps(60, end) infinite,
+      blink-caret .5s step-end infinite;
+}
+  
+/* The typing effect */
+@keyframes typing {
+    from, to { width: 0 }
+    50% { width: 100% }
+    70% { width: 100% }
+    95% { width: 0 }
+}
+  
+/* The typewriter cursor effect */
+@keyframes blink-caret {
+    from, to { border-color: transparent }
+    50% { border-color: orange }
+}
+
+.button-wrapper {
+    height: 30%;
+    width: 100vw;
+    display:flex;
+    justify-content: center;
+    align-items: center;
+}
+
+.game-button {
+    background: transparent;
+    display:flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    height: 100%;
+    width: 15vw;
+
+    margin-left: 4rem;
+    margin-right: 4rem;
+
+    cursor:pointer;
+    background-color: rgba(255, 255, 255, 0.781);
+    border-radius: 10px;
+    box-shadow: 0 0 10px rgba(255, 255, 255, 0.3);
+    backdrop-filter: blur(5px);
+
+    transition: all 0.3s ease-in-out;
+}
+
+.game-button:hover{
+    box-shadow: 0px 10px 20px rgba(0, 0, 0, 0.3);
+}
+
+.game-text{
+    font-family: 'Roboto', sans-serif;
+    font-size: 2rem;
+    text-align: center;
+}
+
+#help-button {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+
+    position: absolute; 
+    top: 2rem;
+    right: 2rem;
+
+    width: 3rem;
+    height: 3rem;
+
+    color: white;
+    background: transparent;
+    border: 2px solid white;
+    border-radius: 50%;
+
+    text-align: right;
+    text-decoration: none;
+    font-size: 2rem;
+    cursor: pointer;
+
+    transition: all 0.5s ease-in-out;
+}
+
+#help-button:hover{
+    background-color: rgba(255, 255, 255, 0.4);
+    box-shadow: 0 0 30px white;
+}
+
+#help-popup {
+    display: none;
+    position: fixed;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    border-radius: 10px;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+    padding: 20px;
+
+    background-color: rgba(255, 255, 255, 0.95); /* set a semi-transparent white background */
+    backdrop-filter: blur(10px); /* apply a blur effect */
+    opacity: 0.95; /* adjust opacity as desired */
+
+    box-shadow: 0px 0px 10px white;
+
+    color: black;
+    z-index: 1000;
+}
+
+#help-popup-text{
+    font-family: 'Roboto', sans-serif;
+    font-size: 1.25rem;
+    text-align: center;
+    margin-bottom: 1rem;
+}
+
+#close-help-button {
+    font-size: 24px;
+    background-color: #9344a9;
+    color: white;
+    padding: 8px 16px;
+    border: none;
+    cursor: pointer;
+    border-radius: 4px;
+    float: left;
+}
+
+#audio-play-button{
+    display:none;
+    justify-content: center;
+    align-items: center;
+    position: absolute;
+    top: 2rem;
+    right: 6rem;
+    height: 3rem;
+    width: 3rem;
+    font-size: larger;
+    cursor: pointer;
+    background:transparent;
+    color: white;
+    border-radius: 50%;
+    border: 2px solid white;
+    transition: all 0.5s ease-in-out;
+}
+
+#audio-play-button:hover{
+    background-color: rgba(255, 255, 255, 0.4);
+    box-shadow: 0 0 30px white;
+}
+
+#game{
+    display: none;
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);    
+    box-shadow: 0 0 50px 10px white;
+}
+
+
+#pause-menu {
+    display: none;
+    /* below the help panel */
+    z-index: 500;
+    font-size: 60px;
+    position: fixed;
+    top: 50%;
+    left: 50%;
+
+    border-radius: 10px;
+
+    transform: translate(-50%, -50%);
+    padding: 20px;
+    text-align: center;
+
+    background-color: rgba(255, 255, 255, 0.95); /* set a semi-transparent white background */
+    backdrop-filter: blur(10px); /* apply a blur effect */
+    opacity: 0.95; /* adjust opacity as desired */
+
+    box-shadow: 0px 0px 10px white;
+}
+  
+#resume-button {
+    font-size: 24px;
+    background-color: #9344a9;
+    color: white;
+    padding: 8px 16px;
+    border: none;
+    cursor: pointer;
+    border-radius: 4px;
+    float: left;
+}
+  
+#exit-button {
+    font-size: 24px;
+    background-color: #9344a9;
+    color: white;
+    padding: 8px 16px;
+    border: none;
+    cursor: pointer;
+    border-radius: 4px;
+    float: right;
+}
+  
+
