--- snake.css
@@ -1,62 +1,0 @@
-html, body {
-    height: 100%;
-    margin: 0;
-  }
-
-  body {
-    background: black;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-  }
-  canvas {
-    border: 1px solid white;
-}
-.play-button {
-  width: 50px;
-  height: 50px;
-  border: none;
-  border-radius: 50%;
-  background-color: #4a4a4a;
-  position: relative;
-  overflow: hidden;
-  cursor: pointer;
-  position: absolute; 
-  top: 1rem;
-  right: 4rem;
-}
-#help-button {
-  background-color: #af4caa;
-  border: none;
-  color: white;
-  padding: 12px 12px;
-  text-align: right;
-  text-decoration: none;
-  display: inline-block;
-  font-size: 16px;
-  margin: 4px 2px;
-  cursor: pointer;
-  position: absolute; 
-  top: 1rem;
-  right: 1rem;
-}
-#help-popup {
-  display: none;
-  position: fixed;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  padding: 20px;
-  background-color: white;
-  box-shadow: 0px 0px 10px rgba(0, 0, 0, 0.2);
-  }
-
-#close-button {
-  background-color: #ae76bd;
-  color: white;
-  padding: 8px 16px;
-  border: none;
-  cursor: pointer;
-  border-radius: 4px;
-  float: right;
-}
