--- tetris.html
@@ -1,41 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <title>Basic Tetris HTML Game</title>
-  <meta charset="UTF-8">
-  <link rel="stylesheet" href="snake.css">
-</head>
-<body>
-<canvas width="320" height="640" id="game"></canvas>
-<button id="help-button">?</button>
-<div id="help-popup">
-    <p>How to play:</p>
-    <p>Move
-      <ul>
-        bring down blocks from the top of the screen 
-      </ul>
-      <ul>
-        You can move the blocks around, either left to right and/or you can rotate them.
-      </ul>
-    </p>
-    <p> Next Piece Preview 
-      <ul>Show the next piece (or pieces) that will enter the playfield. </ul>
-    </p>  
-    <p> Hard Drop 
-      <ul>When the Space key is pressed, the piece should be placed as far down as possible.</ul>
-    </p>  
-    <button id="close-button">Close</button>
-  </div>
-<button id="play" class = "play-button">🔊</button>
-<script>
-  let play = document.getElementById("play");
-  function playMusic() {
-    let audio = new Audio("tetris-sound.mp3");
-    audio.play();
-    audio.loop();
-  }
-  play.addEventListener("click", playMusic);
-</script>
-<script src="tetris.js"></script>
-</body>
-</html>
